init: add filter form
Allows a business post a listing using a form!

Key design choice:
For v.0.0.1 we don't want to handle images ourselves. Instead links the
the desired images are stored instead.

Also: adds tailwindcss and other svelte form libraries. The extra svelte
libraries should be uninstalled once we decide on which is best.